requirements.txt
openai==0.27.8
colorama==0.4.6
python-dotenv==1.0.0
